Moved some more frequently used NVI components
