+++ src/components/nvi/table/rows/NviAdminReportingStatusRow.tsx
@@ -1,9 +1,9 @@
 import { TableCell, TableRow } from '@mui/material';
 import { useTranslation } from 'react-i18next';
-import { InstitutionReport } from '../../../types/nvi.types';
-import { PercentageWithIcon } from '../../_molecules/PercentageWithIcon';
-import { HorizontalBox } from '../../styled/Wrappers';
-import { CenteredTableCell } from '../../tables/table-styles';
+import { InstitutionReport } from '../../../../types/nvi.types';
+import { PercentageWithIcon } from '../../../_molecules/PercentageWithIcon';
+import { HorizontalBox } from '../../../styled/Wrappers';
+import { CenteredTableCell } from '../../../tables/table-styles';
 
 import {
   getNviApprovedCount,
@@ -13,7 +13,7 @@ import {
   getNviSectorLabel,
   getNviTotalCount,
   getPercentageControlled,
-} from './utils/nvi-admin-aggregations-helpers';
+} from '../utils/nvi-admin-aggregations-helpers';
 
 interface NviAdminStatusPageRowProps {
   report: InstitutionReport;
