+++ src/components/nvi/table/rows/NviPublicationPointsRow.tsx
@@ -1,19 +1,19 @@
 import { Link } from '@mui/material';
 import { useState } from 'react';
 import { Link as RouterLink } from 'react-router';
-import { NviCandidateGlobalStatusEnum, NviCandidateStatusEnum } from '../../../api/searchApi';
-import { NviInstitutionStatusResponse } from '../../../types/nvi.types';
-import { Organization } from '../../../types/organization.types';
-import { dataTestId } from '../../../utils/dataTestIds';
-import { getIdentifierFromId } from '../../../utils/general-helpers';
-import { useNviCandidatesParams } from '../../../utils/hooks/useNviCandidatesParams';
-import { getNviCandidatesSearchPath } from '../../../utils/urlPaths';
-import { PercentageWithIcon } from '../../_molecules/PercentageWithIcon';
-import { HorizontalBox } from '../../styled/Wrappers';
-import { CenteredTableCell, TableNumberSkeleton } from '../../tables/table-styles';
+import { NviCandidateGlobalStatusEnum, NviCandidateStatusEnum } from '../../../../api/searchApi';
+import { NviInstitutionStatusResponse } from '../../../../types/nvi.types';
+import { Organization } from '../../../../types/organization.types';
+import { dataTestId } from '../../../../utils/dataTestIds';
+import { getIdentifierFromId } from '../../../../utils/general-helpers';
+import { useNviCandidatesParams } from '../../../../utils/hooks/useNviCandidatesParams';
+import { getNviCandidatesSearchPath } from '../../../../utils/urlPaths';
+import { PercentageWithIcon } from '../../../_molecules/PercentageWithIcon';
+import { HorizontalBox } from '../../../styled/Wrappers';
+import { CenteredTableCell, TableNumberSkeleton } from '../../../tables/table-styles';
+import { ALTERNATIVE_TEXT_INSTEAD_OF_ZERO } from '../utils/constants';
+import { selfOrDescendantHasPointValues } from '../utils/nvi-curator-aggregations-helpers';
 import { NviRowWrapper } from './NviRowWrapper';
-import { ALTERNATIVE_TEXT_INSTEAD_OF_ZERO } from './utils/constants';
-import { selfOrDescendantHasPointValues } from './utils/nvi-curator-aggregations-helpers';
 
 interface NviPublicationPointsTableRowProps {
   organization: Organization;
