+++ src/components/nvi/table/rows/NviRowWrapper.tsx
@@ -3,8 +3,8 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { IconButton, Skeleton, styled, TableCell, TableRow } from '@mui/material';
 import { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Organization } from '../../../types/organization.types';
-import { getLanguageString } from '../../../utils/translation-helpers';
+import { Organization } from '../../../../types/organization.types';
+import { getLanguageString } from '../../../../utils/translation-helpers';
 
 export const StyledSkeleton = styled(Skeleton)({
   width: '2ch',
