+++ src/pages/basic-data/nvi/status/NviAdminReportingStatusPage.tsx
@@ -10,22 +10,22 @@ import {
   Typography,
 } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
-import { NviAdminReportingStatusRow } from '../../../../components/nvi/table/NviAdminReportingStatusRow';
+import { NviPageLayout } from '../../../../components/nvi/NviPageLayout';
+import { NviAdminTableSortSelector } from '../../../../components/nvi/table/NviAdminTableSortSelector';
+import { NviAdminReportingStatusRow } from '../../../../components/nvi/table/rows/NviAdminReportingStatusRow';
 import { TableSkeleton } from '../../../../components/skeletons/TableSkeleton';
 import { NviAdminSortSelectorType } from '../../../../components/sort-selectors/sort-nvi-table/nvi-admin-sort-types';
 import { VerticalBox } from '../../../../components/styled/Wrappers';
 import { CenteredTableCell } from '../../../../components/tables/table-styles';
 import { useInstitutionReportsFilteredAndSortedByUrl } from '../../../../hooks/nvi/useInstitutionReportsFilteredAndSortedByUrl';
 import { InstitutionReport } from '../../../../types/nvi.types';
-import { NviAdminSortSelector } from '../../../basic_data/app_admin/nviAdmin/nviAdminSortSelector/NviAdminSortSelector';
-import { NviStatusWrapper } from '../../../messages/components/NviStatusWrapper';
 
-export const NviAdminStatusPage = () => {
+export const NviAdminReportingStatusPage = () => {
   const { t } = useTranslation();
   const { sortedAndFilteredData, isPending, isError } = useInstitutionReportsFilteredAndSortedByUrl();
 
   return (
-    <NviStatusWrapper
+    <NviPageLayout
       headline={t('basic_data.nvi.reporting_status')}
       topView={
         <Box sx={{ mb: '1rem' }}>
@@ -45,7 +45,7 @@ export const NviAdminStatusPage = () => {
         <Typography>{t('feedback.error.get_nvi_reports')}</Typography>
       ) : (
         <VerticalBox sx={{ width: '100%' }}>
-          <NviAdminSortSelector type={NviAdminSortSelectorType.Status} />
+          <NviAdminTableSortSelector type={NviAdminSortSelectorType.Status} />
           <TableContainer component={Paper} variant="outlined">
             <Table size="small">
               <TableHead>
@@ -70,6 +70,6 @@ export const NviAdminStatusPage = () => {
           </TableContainer>
         </VerticalBox>
       )}
-    </NviStatusWrapper>
+    </NviPageLayout>
   );
 };
